ou t'es espace ou t'es ?

+++ libft/ft_itoa.c
@@ -6,14 +6,14 @@
 /*   By: jcoutare <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/04/19 12:17:43 by jcoutare          #+#    #+#             */
-/*   Updated: 2017/04/20 19:24:37 by jcoutare         ###   ########.fr       */
+/*   Updated: 2017/09/18 13:36:15 by jcoutare         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdio.h>
 
-static char		*nbtostr(char *str, int lgt, int n, int neg)
+static char		*nbtostr(char *str, int lgt, long long n, int neg)
 {
 	if ((str = malloc(sizeof(char) * (lgt + 1))) == NULL)
 		return (NULL);
@@ -29,12 +29,12 @@ static char		*nbtostr(char *str, int lgt, int n, int neg)
 	return (str);
 }
 
-char			*ft_itoa(int n)
+char			*ft_itoa(long long n)
 {
-	char		*str;
-	int			lgt;
-	int			nb;
-	int			neg;
+	char						*str;
+	int							lgt;
+	unsigned long long			nb;
+	int							neg;
 
 	lgt = 1;
 	neg = 1;
